+++ code-challenges/401/getEdge/get-edge.js
@@ -1,6 +1,8 @@
 'use strict';
 
 function getEdge(graph, array){
+  console.log(graph);
+  console.log(array);
   let counter = 0;
   let current;
   let graphArray = graph.getNodes();
@@ -29,31 +31,27 @@ function getEdge(graph, array){
   return `true $${counter}`;
 }
 
+module.exports = getEdge;
+
+// const Graph = require('../graph/graph');
 
 
-const Graph = require('../graph/graph');
+// let graph = new Graph.Graph();
+// let seattle = graph.addNode('Seattle');
+// let bend = graph.addNode('Bend');
+// let LA = graph.addNode('LA');
+// let NY = graph.addNode('New York');
+// // console.log(seattle);
+// graph.addBiDirectionalEdge(seattle, bend, 22);
+// graph.addBiDirectionalEdge(seattle, NY, 13);
+// graph.addBiDirectionalEdge(NY, LA, 120);
+// graph.addBiDirectionalEdge(bend, NY, 30);
+// graph.addBiDirectionalEdge(bend, LA, 30);
+
+// let cityArr = ['Seattle', 'New York', 'LA', 'Seattle'];
+// let cityArr2 = ['Seattle', 'Bend', 'LA'];
+
+// console.log(getEdge(graph, cityArr));
+// console.log(getEdge(graph, cityArr2));
 
 
-let graph = new Graph.Graph();
-let seattle = graph.addNode('Seattle');
-let bend = graph.addNode('Bend');
-let LA = graph.addNode('LA');
-let NY = graph.addNode('New York');
-// console.log(seattle);
-graph.addBiDirectionalEdge(seattle, bend, 22);
-graph.addBiDirectionalEdge(seattle, NY, 13);
-graph.addBiDirectionalEdge(NY, LA, 120);
-graph.addBiDirectionalEdge(bend, NY, 30);
-graph.addBiDirectionalEdge(bend, LA, 30);
-
-let cityArr = ['Seattle', 'New York', 'LA', 'Seattle'];
-let cityArr2 = ['Seattle', 'Bend', 'LA'];
-
-console.log(getEdge(graph, cityArr));
-console.log(getEdge(graph, cityArr2));
-
-// getEdge(graph, cityArr2);
-
-// getEdge(graph, path){
-
-// }
