+++ code-challenges/401/graph/graph.js
@@ -1,6 +1,5 @@
 'use strict';
 
-// const util = require('util');
 
 class Vertex {
   constructor(value){
@@ -46,36 +45,6 @@ class Graph {
     this.addEdge(vertex_b, vertex_a, weight);
   }
 
-  // getNodes(startVertex){
-  //   const stack = [];
-  //   const visitedVertices = new Set();
-  //   const parentPath = new Array();
-
-  //   stack.push(startVertex);
-  //   visitedVertices.add(startVertex);
-
-  //   while(stack.length){
-  //     const currentVertex = stack.pop();
-
-  //     console.log(currentVertex);
-
-  //     const neighbors = this.getNeighbors(currentVertex);
-
-  //     for(let edge of neighbors){
-  //       const neighborVertex = edge.vertex;
-
-  //       if(visitedVertices.has(neighborVertex)){
-  //         continue;
-  //       }else{
-  //         visitedVertices.add(neighborVertex);
-  //       }
-
-  //       stack.push(neighborVertex);
-  //       parentPath.push(neighborVertex, currentVertex);
-  //     }
-  //   }
-  //   return parentPath;
-  // }
   getNodes() {
     if(this.numberOfNodes !== 0){
       return [...this._adjacencyList.keys()];
